requirements.txt
# =============================================================================
# Core Dependencies
# =============================================================================
fastapi
uvicorn[standard]
requests
huggingface_hub
pillow
ddgs
pytz
numpy
pandas
colorama

# =============================================================================
# LLM Frameworks
# =============================================================================
langchain
langgraph
langchain-core
langchain-google-genai
langchain-huggingface
langchain-community
langchain_ollama
langchain-anthropic

# =============================================================================
# Web & Research Tools
# =============================================================================
wikipedia
arxiv
youtube-transcript-api
pytube
markdownify
beautifulsoup4

# =============================================================================
# Document Processing
# =============================================================================
pypdf
pymupdf
python-docx
docx2pdf

# =============================================================================
# Image Processing
# =============================================================================
pillow-heif

# =============================================================================
# Audio/Video Processing
# =============================================================================
speechrecognition
pydub
moviepy

# =============================================================================
# OCR (Optional - requires Tesseract installed)
# =============================================================================
pytesseract

# =============================================================================
# MCP (Model Context Protocol)
# =============================================================================
langchain-mcp-adapters

# =============================================================================
# Observability
# =============================================================================
langfuse

# =============================================================================
# Desktop App (pywebview + pystray)
# =============================================================================
pywebview>=4.0
pystray>=0.19
pyinstaller>=6.0
